+++ packages/common-test-utils/src/presets/pods-core/json.ts
@@ -1,15 +1,20 @@
 import { NotePropsV2 } from "@dendronhq/common-all";
-import { FileTestUtils, writeYAML } from "@dendronhq/common-server";
+import { tmpDir, writeYAML } from "@dendronhq/common-server";
 import fs from "fs-extra";
 import _ from "lodash";
 import path from "path";
-import { EngineTestUtilsV2, NodeTestPresetsV2, NodeTestUtilsV2 } from "../..";
+import {
+  EngineTestUtilsV2,
+  FileTestUtils,
+  NodeTestPresetsV2,
+  NodeTestUtilsV2,
+} from "../..";
 import { TestPresetEntry } from "../../utils";
 
 const IMPORT_BASIC = new TestPresetEntry({
   label: "basic",
   before: async (opts: { wsRoot?: string }) => {
-    const importDir = FileTestUtils.tmpDir().name;
+    const importDir = tmpDir().name;
     const importSrc = path.join(importDir, "import.json");
     const jsonEntries = [
       {