+++ packages/plugin-core/src/test/suite-integ/PublishCommand.test.ts
@@ -1,4 +1,4 @@
-import { DirResult, FileTestUtils } from "@dendronhq/common-server";
+import { DirResult, tmpDir } from "@dendronhq/common-server";
 import * as assert from "assert";
 import { NodeTestPresetsV2 } from "@dendronhq/common-test-utils";
 import { afterEach, beforeEach } from "mocha";
@@ -20,7 +20,7 @@ suite("Publish", function () {
   this.timeout(TIMEOUT);
 
   beforeEach(function () {
-    root = FileTestUtils.tmpDir();
+    root = tmpDir();
     ctx = VSCodeUtils.getOrCreateMockContext();
     DendronWorkspace.getOrCreate(ctx);
   });
@@ -72,7 +72,7 @@ suite("PublishV2", function () {
   this.timeout(TIMEOUT);
 
   beforeEach(function () {
-    root = FileTestUtils.tmpDir();
+    root = tmpDir();
     ctx = VSCodeUtils.getOrCreateMockContext();
     DendronWorkspace.getOrCreate(ctx);
   });