+++ packages/plugin-core/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.2-alpha.4](https://github.com/dendronhq/dendron/compare/v0.14.2-alpha.3...v0.14.2-alpha.4) (2020-11-08)
+
+**Note:** Version bump only for package @dendronhq/plugin-core
+
+## [0.14.2-alpha.3](https://github.com/dendronhq/dendron/compare/v0.14.2-alpha.2...v0.14.2-alpha.3) (2020-11-08)
+
+**Note:** Version bump only for package @dendronhq/plugin-core
+
+## [0.14.2-alpha.2](https://github.com/dendronhq/dendron/compare/v0.14.2-alpha.1...v0.14.2-alpha.2) (2020-11-08)
+
+**Note:** Version bump only for package @dendronhq/plugin-core
+
 ## [0.14.2-alpha.1](https://github.com/dendronhq/dendron/compare/v0.14.2-alpha.0...v0.14.2-alpha.1) (2020-11-07)
 
 **Note:** Version bump only for package @dendronhq/plugin-core